Move creation of tests to render.ts, added more typings
